restApp/requirements.txt
requests == 2.31.0
django == 4.2.4
djangorestframework~=3.14.0
django-cors-headers
pylint
Pillow